[ Refactor ] rename stepper into Multi step registration form
